+++ src/Control/Algebra.hs
@@ -31,7 +31,7 @@ import Control.Effect.State.Internal
 import Control.Effect.Sum ((:+:)(..), Member(..), Members)
 import Control.Effect.Throw.Internal
 import Control.Effect.Writer.Internal
-import Control.Monad ((<=<), join)
+import Control.Monad ((<=<))
 import Data.Functor.Identity
 import qualified Control.Monad.Trans.Except as Except
 import qualified Control.Monad.Trans.Identity as Identity
@@ -108,10 +108,10 @@ handleCoercible = handleIdentity coerce
 -- base
 
 instance Algebra (Lift IO) IO where
-  alg = join . unLift
+  alg (LiftWith with k) = with (Identity ()) coerce >>= k . runIdentity
 
 instance Algebra (Lift Identity) Identity where
-  alg = join . unLift
+  alg (LiftWith with k) = with (Identity ()) coerce >>= k . runIdentity
 
 instance Algebra Choose NonEmpty where
   alg (Choose m) = m True S.<> m False