+++ src/Control/Carrier/Trace/Printing.hs
@@ -18,7 +18,6 @@ import Control.Monad (MonadPlus(..))
 import qualified Control.Monad.Fail as Fail
 import Control.Monad.Fix
 import Control.Monad.IO.Class
-import Control.Monad.IO.Unlift
 import Control.Monad.Trans.Class
 import System.IO
 
@@ -42,12 +41,6 @@ newtype TraceC m a = TraceC (m a)
 instance MonadTrans TraceC where
   lift = TraceC
   {-# INLINE lift #-}
-
-instance MonadUnliftIO m => MonadUnliftIO (TraceC m) where
-  askUnliftIO = TraceC $ withUnliftIO $ \u -> return (UnliftIO (unliftIO u . runTrace))
-  {-# INLINE askUnliftIO #-}
-  withRunInIO inner = TraceC $ withRunInIO $ \run -> inner (run . runTrace)
-  {-# INLINE withRunInIO #-}
 
 instance (MonadIO m, Algebra sig m, Effect c sig) => Algebra (Trace :+: sig) (TraceC m) where
   alg (L (Trace s k)) = liftIO (hPutStrLn stderr s) *> k