+++ src/Control/Effect/Catch/Internal.hs
@@ -14,4 +14,4 @@ data Catch e m k
 deriving instance Functor m => Functor (Catch e m)
 
 instance Effect Functor (Catch e) where
-  thread state handler (Catch m h k) = Catch (handler (m <$ state)) (handler . (<$ state) . h) (handler . fmap k)
+  thread ctx handler (Catch m h k) = Catch (handler (m <$ ctx)) (handler . (<$ ctx) . h) (handler . fmap k)