+++ src/Control/Effect/Cull.hs
@@ -30,7 +30,7 @@ data Cull m k
 deriving instance Functor m => Functor (Cull m)
 
 instance Effect Functor Cull where
-  thread state handler (Cull m k) = Cull (handler (m <$ state)) (handler . fmap k)
+  thread ctx handler (Cull m k) = Cull (handler (m <$ ctx)) (handler . fmap k)
   {-# INLINE thread #-}
 
 -- | Cull nondeterminism in the argument, returning at most one result.