+++ src/Control/Effect/Cut.hs
@@ -36,8 +36,8 @@ data Cut m k
 deriving instance Functor m => Functor (Cut m)
 
 instance Effect Functor Cut where
-  thread _     _       Cutfail    = Cutfail
-  thread state handler (Call m k) = Call (handler (m <$ state)) (handler . fmap k)
+  thread _   _       Cutfail    = Cutfail
+  thread ctx handler (Call m k) = Call (handler (m <$ ctx)) (handler . fmap k)
   {-# INLINE thread #-}
 
 -- | Fail the current branch, and prevent backtracking within the nearest enclosing 'call' (if any).